+++ src/EventHandlers/AppliesEvents.php
@@ -3,21 +3,26 @@
 namespace Spatie\EventSourcing\EventHandlers;
 
 use Spatie\EventSourcing\Handlers;
+use Spatie\EventSourcing\StoredEvents\ShouldBeStored;
 use Spatie\EventSourcing\StoredEvents\StoredEvent;
 
 trait AppliesEvents
 {
-    protected function apply(StoredEvent ...$storedEvents): void
+    protected function apply(StoredEvent | ShouldBeStored ...$storedEvents): void
     {
         foreach ($storedEvents as $storedEvent) {
             $this->applyStoredEvent($storedEvent);
         }
     }
 
-    private function applyStoredEvent(StoredEvent $storedEvent)
+    private function applyStoredEvent(StoredEvent | ShouldBeStored $event)
     {
-        Handlers::find($storedEvent->event, $this)->each(
-            fn (string $handler) => $this->{$handler}($storedEvent->event)
+        $event = $event instanceof StoredEvent
+            ? $event->event
+            : $event;
+
+        Handlers::find($event, $this)->each(
+            fn (string $handler) => $this->{$handler}($event)
         );
     }
 }